requirements.txt
aiogram==2.22.1
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
colorama==0.4.5
commonmark==0.9.1
Deprecated==1.2.13
distlib==0.3.6
filelock==3.8.0
frozenlist==1.3.1
greenlet==1.1.3.post0
huggingface-hub==0.10.0
HyperPyYAML==1.0.1
identify==2.5.5
idna==3.4
joblib==1.2.0
multidict==6.0.2
nodeenv==1.7.0
numpy==1.21.6
packaging==21.3
pandas==1.3.5
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2-binary==2.9.4
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy==1.7.3
sentencepiece==0.1.97
six==1.16.0
speechbrain==0.5.13
SQLAlchemy==1.4.42
toml==0.10.2
torch==1.12.1
torchaudio==0.12.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.5
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
